Conflicts:
	TeamProjDjango/src/HomePage/models.py
	TeamProjJava/WebContent/css/clean-blog.css
	TeamProjJava/WebContent/header.jsp
	TeamProjJava/WebContent/index.jsp
	TeamProjJava/WebContent/loginorReg.jsp
	TeamProjJava/WebContent/post.jsp
	TeamProjJava/src/common/SqlMapConfig.xml
	TeamProjJava/src/controller/MainControl.java
	TeamProjJava/src/dao/BoardDAO.java
	TeamProjJava/src/dto/Buser.java
	TeamProjJava/src/mapper/Buser.xml
	TeamProjJava/src/service/LoginService.java
	TeamProjJava/src/service/MemberInfoService.java
	TeamProjJava/src/service/MemberJoinService.java
	TeamProjJava/target/classes/common/MBUtils.class
	TeamProjJava/target/classes/common/SqlMapConfig.xml
	TeamProjJava/target/classes/controller/MainControl.class
	TeamProjJava/target/classes/dto/Buser.class
	TeamProjJava/target/classes/mapper/Buser.xml
	TeamProjJava/target/classes/service/LoginService.class
	TeamProjJava/target/classes/service/MemberInfoService.class
	TeamProjJava/target/classes/service/MemberJoinService.class
	TeamProjJava/target/m2e-wtp/web-resources/META-INF/maven/TeamProjJava/TeamProjJava/pom.properties
	TeamProjPHP/htdocs/index.php
	TeamProjPHP/htdocs/login/login.php

+++ TeamProjPHP/htdocs/member/member_delete_action.php
@@ -19,20 +19,27 @@
     } else {
         die("연결 실패 : " .mysqli_error());
     }
-    $sql = "DELETE FROM BUSER WHERE id='"
-        . $user_id . "' AND pw=" . $user_pw."";
-    $result = $conn -> prepare($sql);
     
-    $result -> execute();
-    if($result -> execute() == TRUE){
+    $selsql = $conn -> prepare("SELECT * FROM BUSER WHERE ID = :userid");
+    $selsql -> bindParam(':userid', $user_id, PDO::PARAM_STR);
+    $selsql -> execute();
+    $oci_selsql = $selsql -> fetchAll(PDO::FETCH_ASSOC);
+    
+    
+    if($oci_selsql[0]['ID'] == $user_id && $oci_selsql[0]['PW'] == $user_pw){
+        $sql = "DELETE FROM BUSER WHERE id='"
+            . $user_id . "' AND pw=" . $user_pw."";
+        $result = $conn -> prepare($sql);
+        $result -> execute();
         echo "삭제 성공";
         session_start();
         session_destroy();
+        header("Location: ../index.php"); //헤더 함수를 이용해서 리다이렉션 시킬 수 있다.
     }else {
-        echo "삭제 실패";
+        echo "<script> alert('회원탈퇴에 실패하셨습니다.');</script>";
+        header("Location: ../member/member_delete_form.php"); //헤더 함수를 이용해서 리다이렉션 시킬 수 있다.
     }
     //헤더함수를 이용하여 리스트 페이지로 리다이렉션
-    header("Location: ../index.php"); //헤더 함수를 이용해서 리다이렉션 시킬 수 있다.
 ?> 
 </body>
 </html>
